requirements.txt
flask
mysql-connector-python
psycopg2-binary
marshmallow
Flask-SQLAlchemy
PyJWT
flasgger
python-dotenv